backend/requirements.txt
fastapi
uvicorn
pandas
requests
fpdf2
python-dotenv